Reorganize engine modules under src/engine/
Collapse src/kernel/, src/ingest/, and the engine-portion of src/query/
into a single src/engine/ module with {bitmap, primary_dir, tables,
ingest, query} submodules. Move the per-domain query orchestrators
(query/indexed.rs, query/runner.rs) to src/logs/ as indexed_query.rs
and scan_query.rs; these construct QueryLogsResponse via LogMaterializer
and are not engine code.

Pure file moves and mod/use rewrites; no logic changes. The engine
module still contains log-specific leakage (stream_entries_for_log,
IndexedLogClause) which subsequent commits in the generalization stack
clean out.
