+++ monad-chain-data/src/engine/ingest/bitmap_compaction.rs
@@ -16,14 +16,14 @@
 use std::collections::{BTreeMap, BTreeSet};
 
 use crate::{
-    error::Result,
-    kernel::{
+    engine::{
         bitmap::{
             compact_bitmap_page, global_page_start, local_page_start, stream_page_global_start,
             BitmapFragmentWrite,
         },
         tables::LogTables,
     },
+    error::Result,
     store::{BlobStore, MetaStore},
 };
 
@@ -176,7 +176,7 @@ mod tests {
     use bytes::Bytes;
 
     use super::{build_compaction_plan, touched_streams_by_page, BitmapCompactionPlan};
-    use crate::kernel::bitmap::{BitmapFragmentWrite, STREAM_PAGE_LOCAL_ID_SPAN};
+    use crate::engine::bitmap::{BitmapFragmentWrite, STREAM_PAGE_LOCAL_ID_SPAN};
 
     #[test]
     fn compaction_plan_carries_previous_open_streams_into_the_sealed_start_page() {
