+++ monad-chain-data/src/engine/query/mod.rs
@@ -13,6 +13,6 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-pub mod bitmap;
-pub mod primary_dir;
-pub mod tables;
+pub(crate) mod bitmap;
+pub(crate) mod directory_resolver;
+pub(crate) mod window;
