+++ monad-chain-data/src/lib.rs
@@ -15,21 +15,19 @@
 
 pub mod api;
 pub mod blocks;
+pub mod engine;
 pub mod error;
 pub mod family;
-pub mod ingest;
-pub mod kernel;
 pub mod logs;
 pub mod primitives;
-pub mod query;
 pub mod store;
 
 pub use alloy_primitives::{Address, Bytes, Log, LogData, B256};
 pub use api::{IngestOutcome, MonadChainDataService};
 pub use blocks::{Block, QueryBlocksRequest, QueryBlocksResponse};
+pub use engine::tables::Tables;
 pub use error::MonadChainDataError;
 pub use family::{FinalizedBlock, Hash32};
-pub use kernel::tables::Tables;
 pub use logs::{LogEntry, LogFilter, LogsRelations, QueryLogsRequest, QueryLogsResponse};
 pub use primitives::{
     limits::{LimitExceededKind, QueryEnvelope, QueryLimits},
