+++ monad-chain-data/src/logs/indexed_query.rs
@@ -15,13 +15,15 @@
 
 use roaring::RoaringBitmap;
 
-use super::{
-    bitmap::load_clause_bitmap_for_shard, directory_resolver::LogIdResolver,
-    window::resolve_log_window,
-};
 use crate::{
+    engine::{
+        query::{
+            bitmap::load_clause_bitmap_for_shard, directory_resolver::LogIdResolver,
+            window::resolve_log_window,
+        },
+        tables::Tables,
+    },
     error::{MonadChainDataError, Result},
-    kernel::tables::Tables,
     logs::{LogMaterializer, QueryLogsRequest, QueryLogsResponse},
     primitives::{page::QueryOrder, range::ResolvedBlockWindow, refs::BlockSpan, state::LogId},
     store::{BlobStore, MetaStore},
